test: move new test to old file
